requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-filter==2.4.0
django-recaptcha==2.0.6
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
mysqlclient==1.4.6
oauthlib==3.1.0
psycopg2==2.8.6
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
